backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
python-multipart==0.0.6
pandas==2.1.3
openpyxl==3.1.2
python-jose==3.3.0
passlib==1.7.4
python-dotenv==1.0.0
aiosqlite==0.19.0
bcrypt==4.0.1
pydantic-settings==2.1.0
pydantic==2.5.2
email-validator==2.1.0.post1 